backend/requirements.txt
bottle==
google.cloud==